+++ src/main/java/core/mvc/tobe/AnnotationHandlerMapping.java
@@ -1,18 +1,37 @@
 package core.mvc.tobe;
 
 import com.google.common.collect.Maps;
+import core.annotation.web.Controller;
+import core.annotation.web.RequestMapping;
 import core.annotation.web.RequestMethod;
 import core.di.factory.BeanFactory;
-import core.di.factory.BeanScanner;
 import core.mvc.HandlerMapping;
+import core.mvc.tobe.support.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.core.LocalVariableTableParameterNameDiscoverer;
+import org.springframework.core.ParameterNameDiscoverer;
 
 import javax.servlet.http.HttpServletRequest;
-import java.util.Map;
+import java.lang.reflect.Method;
+import java.util.*;
+import java.util.stream.Collectors;
+
+import static java.util.Arrays.asList;
 
 public class AnnotationHandlerMapping implements HandlerMapping {
     private static final Logger logger = LoggerFactory.getLogger(AnnotationHandlerMapping.class);
+
+    private static final List<ArgumentResolver> argumentResolvers = asList(
+            new HttpRequestArgumentResolver(),
+            new HttpResponseArgumentResolver(),
+            new RequestParamArgumentResolver(),
+            new PathVariableArgumentResolver(),
+            new ModelArgumentResolver()
+    );
+
+    private static final ParameterNameDiscoverer nameDiscoverer = new LocalVariableTableParameterNameDiscoverer();
+
 
     private Map<HandlerKey, HandlerExecution> handlerExecutions = Maps.newHashMap();
 
@@ -21,7 +40,7 @@ public class AnnotationHandlerMapping implements HandlerMapping {
 
     public void initialize() {
         logger.info("## Initialized Annotation Handler Mapping");
-        handlerExecutions.putAll(BeanScanner.getInstance().getHandlers(BeanFactory.getInstance().getBeans()));
+        handlerExecutions.putAll(this.getHandlers(BeanFactory.getInstance().getBeans()));
     }
 
     public Object getHandler(HttpServletRequest request) {
@@ -29,6 +48,38 @@ public class AnnotationHandlerMapping implements HandlerMapping {
         RequestMethod rm = RequestMethod.valueOf(request.getMethod().toUpperCase());
         logger.debug("requestUri : {}, requestMethod : {}", requestUri, rm);
         return getHandlerInternal(new HandlerKey(requestUri, rm));
+    }
+
+    public Map<HandlerKey, HandlerExecution> getHandlers(Map<Class<?>, Object> beans) {
+
+        Map<HandlerKey, HandlerExecution> handlers = new HashMap<>();
+
+        Set<Object> controllers = beans.values().stream()
+                .filter(this::isController)
+                .collect(Collectors.toSet());
+
+        for (Object controller : controllers) {
+            addHandlerExecution(handlers, controller, controller.getClass().getMethods());
+        }
+
+        return handlers;
+    }
+
+    private boolean isController(Object clazz) {
+        return Arrays.stream(clazz.getClass().getDeclaredAnnotations())
+                .anyMatch(annotation -> annotation.annotationType().equals(Controller.class));
+    }
+
+    private void addHandlerExecution(Map<HandlerKey, HandlerExecution> handlers, final Object target, Method[] methods) {
+        Arrays.stream(methods)
+                .filter(method -> method.isAnnotationPresent(RequestMapping.class))
+                .forEach(method -> {
+                    RequestMapping requestMapping = method.getAnnotation(RequestMapping.class);
+                    HandlerKey handlerKey = new HandlerKey(requestMapping.value(), requestMapping.method());
+                    HandlerExecution handlerExecution = new HandlerExecution(nameDiscoverer, argumentResolvers, target, method);
+                    handlers.put(handlerKey, handlerExecution);
+                    logger.info("Add - method: {}, path: {}, HandlerExecution: {}", requestMapping.method(), requestMapping.value(), method.getName());
+                });
     }
 
     private HandlerExecution getHandlerInternal(HandlerKey requestHandlerKey) {
@@ -40,4 +91,5 @@ public class AnnotationHandlerMapping implements HandlerMapping {
 
         return null;
     }
+
 }
