+++ src/test/java/core/di/factory/BeanScannerTest.java
@@ -4,12 +4,12 @@ import core.di.factory.example.JdbcQuestionRepository;
 import core.di.factory.example.JdbcUserRepository;
 import core.di.factory.example.MyQnaService;
 import core.di.factory.example.QnaController;
-import core.mvc.tobe.HandlerExecution;
-import core.mvc.tobe.HandlerKey;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
 
 class BeanScannerTest {
 
@@ -21,14 +21,4 @@ class BeanScannerTest {
         Assertions.assertThat(preInstantiatedBeans).isEqualTo(expected);
     }
 
-    @Test
-    void getHandlersTest() {
-        QnaController qnaController = new QnaController(new MyQnaService(new JdbcUserRepository(), new JdbcQuestionRepository()));
-        Map<Class<?>, Object> beans = new HashMap<>() {{
-            put(QnaController.class, qnaController);
-        }};
-        Map<HandlerKey, HandlerExecution> handlers = BeanScanner.getInstance().getHandlers(beans);
-
-        Assertions.assertThat(handlers.size()).isEqualTo(1);
-    }
 }
