+++ src/test/java/core/mvc/tobe/AnnotationHandlerMappingTest.java
@@ -1,8 +1,13 @@
 package core.mvc.tobe;
 
 import core.di.factory.BeanFactory;
+import core.di.factory.example.JdbcQuestionRepository;
+import core.di.factory.example.JdbcUserRepository;
+import core.di.factory.example.MyQnaService;
+import core.di.factory.example.QnaController;
 import next.dao.UserDao;
 import next.model.User;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
@@ -10,6 +15,8 @@ import org.springframework.mock.web.MockHttpServletResponse;
 import support.test.DBInitializer;
 
 import java.lang.reflect.InvocationTargetException;
+import java.util.HashMap;
+import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -42,6 +49,18 @@ public class AnnotationHandlerMappingTest {
         assertThat(request.getAttribute("user")).isEqualTo(user);
     }
 
+    @Test
+    void getHandlersTest() {
+        QnaController qnaController = new QnaController(new MyQnaService(new JdbcUserRepository(), new JdbcQuestionRepository()));
+        Map<Class<?>, Object> beans = new HashMap<>() {{
+            put(QnaController.class, qnaController);
+        }};
+        Map<HandlerKey, HandlerExecution> handlers = handlerMapping.getHandlers(beans);
+
+        Assertions.assertThat(handlers.size()).isEqualTo(1);
+    }
+
+
     private void createUser(User user) throws Exception {
         MockHttpServletRequest request = new MockHttpServletRequest("POST", "/users");
         request.setParameter("userId", user.getUserId());
@@ -52,4 +71,5 @@ public class AnnotationHandlerMappingTest {
         HandlerExecution execution = (HandlerExecution) handlerMapping.getHandler(request);
         execution.handle(request, response);
     }
+
 }
